Cursed last cypress test
